OFString: Add a cast which is needed on NetBSD

+++ src/OFString.m
@@ -1473,7 +1473,7 @@ static struct {
 		if (options & OF_JSON_REPRESENTATION_IDENTIFIER) {
 			const char *cString = [self UTF8String];
 
-			if ((!isalpha(cString[0]) && cString[0] != '_' &&
+			if ((!isalpha((int)cString[0]) && cString[0] != '_' &&
 			    cString[0] != '$') ||
 			    strpbrk(cString, " \n\r\t\b\f\\\"'") != NULL) {
 				[JSON prependString: @"\""];
